refactor: generateReview の Prompt Compiler ブロックを compiler-stage へ抽出する
`generateReview`（src/lib/review-engine.mjs）が 462 行あり、779 行のファイルの
59% を占めていた。src/cli.mjs のリファクタ（#1901 / #1905）と同じ型で、面ごとに
関数を切り出す第 1 段として Prompt Compiler の配線ブロックを抽出する。

- src/prompt/compiler-stage.mjs を新設し、`runPromptCompilerStage` を置く
- 同ブロックだけが使っていた `promptFingerprint` /
  `buildPromptCompilerObservation` / `PROMPT_HASH_CHARS` も同じ module へ移す
- `generateReview` は 462 行 → 428 行、review-engine.mjs は 779 行 → 688 行

振る舞いは変えていない。`generateReview` のシグネチャは無変更で、tests/ は
1 行も触っていない。
